% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tesseract-package.R
\docType{package}
\name{tesseract-package}
\alias{tesseract-package}
\title{Open Source OCR Engine}
\description{
Bindings to 'Tesseract':
a powerful optical character recognition (OCR) engine that supports over 100
languages. The engine is highly configurable in order to tune the detection
algorithms and obtain the best possible results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/tesseract/ (website) https://github.com/ropensci/tesseract (devel)}
  \item Report bugs at \url{https://github.com/ropensci/tesseract/issues}
}

}
\author{
\strong{Maintainer}: Jeroen Ooms \email{jeroen@berkeley.edu} (\href{https://orcid.org/0000-0002-4035-0289}{ORCID})

Authors:
\itemize{
  \item Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})
}

}
