% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\name{ocr}
\alias{ocr}
\alias{tesseract}
\title{Tesseract OCR}
\usage{
ocr(image, format)
}
\arguments{
\item{image}{file path or raw vector to image}

\item{format}{must be one of "png", "jpeg", "tiff", "gif"}
}
\description{
Extract text from an image file.
}
\examples{
# Some packages
library(pdftools)
library(tesseract)
library(png)
library(jpeg)
library(tiff)

# Render pdf to png
setwd(tempdir())
news <- file.path(Sys.getenv("R_DOC_DIR"), "NEWS.pdf")
bitmap <- pdf_render_page(news, dpi = 300)
png::writePNG(bitmap, "page.png")
jpeg::writeJPEG(bitmap, "page.jpg")
tiff::writeTIFF(bitmap, "page.tiff")

# Extract text from images
txt1 <- ocr("page.png")
txt2 <- ocr("page.jpg")
txt3 <- ocr("page.tiff")
cat(txt1)
}

